Seperated video input through webcam and file
