+++ Code/Detection_Through_Video.py
@@ -1,0 +1,22 @@
+import cv2
+import numpy as np
+from Motion_Detection import *
+
+def main():
+    while True:
+        # Capturing video
+        video = cv2.VideoCapture('../pebrin database/Video-7_2019-08-14.wmv')
+        isclosed = 0
+
+        MotionDetection(video)
+
+        if isclosed:
+            break
+
+    video.release()
+
+    # Destroying all the windows
+    cv2.destroyAllWindows()
+
+if __name__== "__main__":
+    main()
