+++ Code/Detection_Through_Webcam.py
@@ -1,0 +1,22 @@
+import cv2
+import numpy as np
+from Motion_Detection import *
+
+def main():
+    while True:
+        # Capturing video
+        video = cv2.VideoCapture(0)
+        isclosed = 0
+
+        MotionDetection(video)
+
+        if isclosed:
+            break
+
+    video.release()
+
+    # Destroying all the windows
+    cv2.destroyAllWindows()
+
+if __name__== "__main__":
+    main()
